refactored protobuf code
refactored protobuf helper methods and constructor into
protobufHelperFunctions.js
