+++ spec/routesSpec.js
@@ -8,6 +8,9 @@ var express = require('express');
 var protoBuf = require('protobufjs');
 var app = server.app;
 
+/* Load protobuf helper methods */
+var pbhelper = require('../protobuf/protobufHelperFunctions')
+
 describe("Routes:", function(done) {
     var authUn;
     var authUnBadDid;
@@ -20,7 +23,6 @@ describe("Routes:", function(done) {
     var protoPrekeys;
 
     it('Create Credentials & Keys for testing', function(done) {
-        //this.timeout(10000);
         request(app)
         .get('/test/axolotl')
         .end(function(err, res) {
@@ -40,23 +42,7 @@ describe("Routes:", function(done) {
     });
 
     it('Create Prekeys protobuf', function(done) {
-        //experiment with protoBuf
-        var builder = protoBuf.loadProtoFile("protobuf/keys.proto"); //appears to automaticly search from root of project
-        var Protoprekeys = builder.build("protoprekeys");
-        var Prekeys = Protoprekeys.Prekeys;
-        var Prekey = Protoprekeys.Prekey;
-        var Keypair = Protoprekeys.Keypair;
-
-        var ConstructKeysProtobuf = function (lastResortKey, prekeys){
-            var protoLastResortKey = new Prekey(Number(lastResortKey.id), new Keypair(ab2str(lastResortKey.keyPair.public), ab2str(lastResortKey.keyPair.private)));
-
-            var protoPrekeys = new Prekeys(protoLastResortKey);
-            for (var i=0; i<prekeys.length; i++) {
-                protoPrekeys['prekeys'][i] = new Prekey(prekeys[i].id, new Keypair(ab2str(prekeys[i].keyPair.public), ab2str(prekeys[i].keyPair.private)));
-            }
-            return protoPrekeys;
-        };
-        protoPrekeys = ConstructKeysProtobuf(lastResortKey, prekeys);
+        protoPrekeys = pbhelper.constructKeysProtobuf(lastResortKey, prekeys);
         done();
     });
 
@@ -201,16 +187,3 @@ describe("Routes:", function(done) {
         done();
     });
 });
-
-var ab2str = function(buf) {
-  return String.fromCharCode.apply(null, new Int8Array(buf));
-};
-
-var str2ab = function(str) {
-  var buf = new ArrayBuffer(str.length*2); // 2 bytes for each char
-  var bufView = new Int8Array(buf);
-  for (var i=0, strLen=str.length; i<strLen; i++) {
-    bufView[i] = str.charCodeAt(i);
-  }
-  return buf;
-};
